Configure `arduino/arduino-lint-action` to Library Manager "update" mode
The `arduino/arduino-lint-action` GitHub Actions action offers a set of checks for the additional requirements specific
to the Arduino Library Manager index.

There are differences in these requirements depending on whether the library is already in the index. By default, these
checks are in the "submit" mode, where the library is checked for compliance with the requirements to be accepted into
the Library Manager index. Once the library has been accepted, the action's configuration must be adjusted to the
"update" mode, where it is checked for the requirements for a new release of a library already in the Library Manager
index to be added.

The library has now been added to the Library Manager index, but the action still had the "submit" configuration, which
resulted in a spurious workflow run failure due to violation of rule LP017. Reconfiguring the action to "update" mode
will cause the action to provide accurate results.